requirements.txt
Flask==3.0.2
Pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0
requests==2.31.0
Werkzeug==3.0.1
beautifulsoup4==4.12.3
opencv-python==4.9.0.80
scipy==1.12.0
setuptools>=69.0.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
torch==2.2.0
torchvision==0.17.0 